Move tracing related code to the Tracing namespace
